+++ android/src/main/java/com/opentokreactnative/utils/Utils.java
@@ -5,10 +5,16 @@ import com.opentok.android.Publisher;
 import com.opentok.android.PublisherKit;
 import com.opentok.android.Subscriber;
 import com.opentok.android.SubscriberKit;
+import com.opentok.android.Session.Builder.TransportPolicy;
+import com.opentok.android.Session.Builder.IncludeServers;
+import com.opentok.android.Session.Builder.IceServer;
 import com.opentokreactnative.OTRN;
+
+import com.facebook.react.bridge.ReadableArray;
 
 import java.util.ArrayList;
 import java.util.Map;
+import java.util.List;
 
 public final class Utils {
 
@@ -58,5 +64,38 @@ public final class Utils {
             }
         }
         return "";
+    }
+
+    public static IncludeServers sanitizeIncludeServer(String value) {
+        IncludeServers includeServers =  IncludeServers.All;
+        if (value != null && value.equals("custom")) {
+            includeServers = IncludeServers.Custom;
+        }
+        return includeServers;
+    }
+
+    public static TransportPolicy sanitizeTransportPolicy(String value) {
+        TransportPolicy transportPolicy =  TransportPolicy.All;
+        if (value != null && value.equals("relay")) {
+            transportPolicy =  TransportPolicy.Relay;
+        }  
+        return transportPolicy;
+    }
+
+    public static List<IceServer> sanitizeIceServer(ReadableArray serverList) {
+        List<IceServer> iceServers = new ArrayList<>();
+        if (serverList != null) {
+            for (int i = 0; i < serverList.size(); i++) {
+                for (int j = 0; j < serverList.getMap(i).getArray("urls").size(); j++) {
+                    iceServers.add(new IceServer(
+                            serverList.getMap(i).getArray("urls").getString(j),
+                            serverList.getMap(i).getString("username"),
+                            serverList.getMap(i).getString("credential")
+
+                    ));
+                }
+            }
+        }
+        return iceServers;
     }
 }