+++ ios/OpenTokReactNative/Utils/Utils.swift
@@ -65,4 +65,46 @@ class Utils {
             return "CodecNotSupported"
         }
     }
+    
+    static func sanitizeIncludeServer(_ value: Any)  -> OTSessionICEIncludeServers {
+        var includeServers = OTSessionICEIncludeServers.all;
+        if let includeServer = value as? String, includeServer == "custom" {
+            includeServers = OTSessionICEIncludeServers.custom;
+        }
+        return includeServers;
+    }
+    
+    static func sanitizeTransportPolicy(_ value: Any)  -> OTSessionICETransportPolicy {
+        var transportPolicy = OTSessionICETransportPolicy.all;
+        if let policy = value as? String, policy == "relay" {
+            transportPolicy = OTSessionICETransportPolicy.relay;
+        }
+        return transportPolicy;
+    }
+    
+    static func sanitiseServerList(_ serverList: Any) -> [(urls: [String], userName: String, credential: String)] {
+        var iceServerList: [([String], String, String)] = []
+        
+        if let serverList = serverList as? [[String: Any]] {
+            for server in serverList {
+                if let urls = server["urls"] as? [String], let username = server["username"] as? String, let credential = server["credential"] as? String {
+                    iceServerList.append((urls, username, credential))
+                }
+            }
+        }
+        return iceServerList
+    }
+    
+    static func sanitizeIceServer(_ serverList: Any, _ transportPolicy: Any, _ includeServer: Any) -> OTSessionICEConfig {
+        let myICEServerConfiguration: OTSessionICEConfig = OTSessionICEConfig();
+        myICEServerConfiguration.includeServers = Utils.sanitizeIncludeServer(includeServer);
+        myICEServerConfiguration.transportPolicy = Utils.sanitizeTransportPolicy(transportPolicy);
+        let serverList = Utils.sanitiseServerList(serverList);
+        for server in serverList {
+            for url in server.urls {
+                myICEServerConfiguration.addICEServer(withURL: url, userName: server.userName, credential: server.credential, error: nil);
+            }
+        }
+        return myICEServerConfiguration;
+    }
 }